# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Aqua API",
    "description": "API documentation for the Aqua track, including users and authentication operations.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local development server"
    },
    {
      "url": "https://aqua-api-fkf8.onrender.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/water": {
      "post": {
        "tags": [
          "Water"
        ],
        "summary": "Creating a record of drinking water",
        "operationId": "createWaterRecord",
        "description": "***Creating a new record of water drunk at a specific time***\n\n**Required fields**\n- date:   *Date and time of drinking water*\n          (формат \"yyyy-mm-ddThh:mm\")\n- volume: *The volume of water drunk at the specified time*\n          (in milliliters, from 50 to 5000)\n\n### Security\n- This endpoint requires authentication via a Bearer Token.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "volume"
                ],
                "properties": {
                  "date": {
                    "type": "string",
                    "description": "Date and time (формат \"yyyy-mm-ddThh:mm\")",
                    "example": "2025-01-10T14:02"
                  },
                  "volume": {
                    "type": "number",
                    "description": "The volume of water drunk (in milliliters)",
                    "example": 300
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a waterRecord!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/water/{id}": {
      "delete": {
        "tags": [
          "Water"
        ],
        "summary": "Deleting a record of drinking water",
        "operationId": "deleteWaterRecord",
        "description": "**Deleting a record of drinking water**\n\nThis endpoint allows you to remove an existing entry from your list.\nOnce deleted, a recording cannot be restored.\n\n### Security\n- This endpoint requires authentication via a Bearer Token.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier for the entry will be removed",
            "example": "6779c35e16fb05e4092b039b"
          }
        ],
        "responses": {
          "204": {
            "description": "deleted successfully **  (No Content)**",
            "content": {}
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Record not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": [
          "Water"
        ],
        "summary": "Change in an existing record of the amount of water you drink",
        "description": "**Update the field for recording the water consumed at a specific point in time by its ID.**\n\nThis endpoint allows you to partially update an existing water drink record.\n\n### Security\n- This endpoint requires authentication via a Bearer Token.\n",
        "operationId": "updateWaterRecord",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the record to update.",
            "example": "6779b9f903d039994c667196"
          }
        ],
        "requestBody": {
          "description": "Field <volume> to update in the water record",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "volume": {
                    "description": "Volume of water drunk",
                    "type": "number",
                    "example": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully patched a water user record!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated the record!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Record not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/water/day/{date}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get a list of posts for the day",
        "operationId": "getWaterDay",
        "description": "** List of records of a user's water consumption at a specific point in time during the day**\n** The entries are sorted by time in ascending order. **\nThis endpoint selects records without service fields\n\n### Security\n- This endpoint requires authentication via a Bearer Token.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "date",
            "example": "2025-01-04"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found records!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully retrieved records."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/waterDay"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/water/month/{month}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get list of records",
        "operationId": "getWaterMonth",
        "description": "** List of days of the month with the amount of water drunk and percentage of the daily norm**\n\n### Security\n  - This endpoint requires authentication via a Bearer Token.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Month for which water information is needed",
            "example": "2025-01"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found contacts!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found records!"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/waterMonth"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use the Bearer token for authorization."
      }
    },
    "schemas": {
      "water": {
        "type": "object",
        "required": [
          "date",
          "volume"
        ],
        "properties": {
          "_id": {
            "description": "Record's id",
            "type": "string",
            "example": "6779b9f903d039994c667196"
          },
          "userId": {
            "description": "ID of the user who owns the record",
            "type": "string",
            "example": "675a2bdc1396640c2c39d34c"
          },
          "date": {
            "description": "Date and time of drinking water",
            "type": "string",
            "example": "2025-01-03T18:02"
          },
          "volume": {
            "description": "Volume of water drunk",
            "type": "number",
            "example": 500
          }
        }
      },
      "waterDay": {
        "type": "array",
        "required": [
          "waterDayByHour"
        ],
        "properties": {
          "day": {
            "description": "Date and time of drinking water",
            "type": "string",
            "example": "2025-01-03"
          },
          "totalVolume": {
            "description": "Amount of water drunk on a given day",
            "type": "number",
            "example": 1500
          },
          "percent": {
            "description": "Percentage of water consumed relative to the daily norm",
            "type": "number",
            "example": 100
          }
        }
      },
      "waterMonth": {
        "type": "array",
        "required": [
          "waterMonthByDay\""
        ],
        "properties": {
          "_id": {
            "description": "Record's id",
            "type": "string",
            "example": "6779b9f903d039994c667196"
          },
          "date": {
            "description": "Date and time of drinking water",
            "type": "string",
            "example": "2025-01-03T18:02"
          },
          "volume": {
            "description": "Volume of water drunk",
            "type": "number",
            "example": 500
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "BadRequestError",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "BadRequestError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "BadRequestError"
                    },
                    "errors": {
                      "type": "array",
                      "example": [
                        {}
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized Error"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "409": {
        "description": "**Conflict Error (409)**\n\nThis error occurs when there is a conflict with the current state of the resource.\n- In this case, the conflict occurs because the specified water consumption time already exists in the database.\n",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "description": "HTTP status code indicating the error type.",
                  "example": 409
                },
                "message": {
                  "type": "string",
                  "description": "A brief summary of the conflict error.",
                  "example": "ConflictError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message",
                    "errors"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Specific details about the error, such as the conflicting field.",
                      "example": "Email already in use"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal server Error"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Detailed error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}